requirements.txt
py-clob-client>=0.17.0
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.7.0
flask>=3.0.0
pywebview>=5.0
pyinstaller>=6.0
